// Bibliotheque_Front/src/app/modules/bibliotheque/models/catalogue.ts
export class Catalogue {
  doc_cote: string | undefined;
  doc_date_edition: any;
  doc_lieu_edition: string | undefined;
  doc_editeur: string | undefined;
  doc_nombre_page: string | undefined;
  doc_sous_titre: string | undefined;
  doc_isbn_issn: string | undefined;
  doc_resume: string | undefined;
  doc_titre: string | undefined;
  doc_volume: string | undefined;
  doc_numero_edition: string | undefined;
  doc_disponibilite: string | undefined;
  theme_code: string | undefined;
  type_code: string | undefined;
  doc_date_saisie: any;
}